Fix syscollector deltas IT - suggested review comments
- Place some common functions/classes to wazuh_testing
- Place some fixtures to analysisd conftest.py
- Fix some typos
